twister: move ZEPHYR_BASE to constants to reduce dependencies
Moving ZEPHYR_BASE from environment.py to constants.py
allows importing classes like HardwareMap in
pytest-twister-harness module without pulling in unnecessary
dependencies from environment.py.

This refactoring improves module independence and makes selective
imports cleaner across the twister Python package structure.
